refactor: import directly from env/event-target
